Half decent working stuff
This whole thing was brought to you by rum and coke.
It seems to work on my end. All local hosted (I gotta get that AWS up) and it's not to complex.
I spent more time making the clock work than anything.
If you wonder why something is done the way it is, change it and see if it breaks. If it doesn't congrads, tell us and we've learned something. If it does congrads, you've learned something.
I tried to comment the whole thing nice and pretty-ish and provide a "but what if I want to do this" for most scenarios. Should be a decent jumping in point.
I don't like pickles they're gross.